requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
Click==7.0
coverage==5.0.3
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-extensions==2.2.8
gunicorn==20.0.4
pathspec==0.7.0
pip-autoremove==0.9.1
psycopg2-binary==2.8.4
python-dotenv==0.11.0
pytz==2019.3
regex==2020.2.18
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.0
whitenoise==5.0.1
